Add support for subprocess debugging
Added options and events to support subprocess debugging.
Implemented experimental subprocess debugging for the Mono
soft debugger. This is done by intercepting calls to start
the debugger and injecting the debugger agent options.
This should ideally be handled by Mono itself, and we can
propose to incorporate the required hooks to Mono, but meanwhile
we can try this hacky implementation.